requirements.txt
Django>=4.2.0
Jinja2>=3.1.2
django-unfold>=0.6.0
psycopg2>=2.9.6
psycopg2-binary>=2.9.6